myhealing/requirements.txt
asgiref==3.5.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==37.0.4
defusedxml==0.7.1
Deprecated==1.2.13
dj-rest-auth==2.2.5
Django==4.0.5
django-allauth==0.51.0
django-filter==22.1
django-oauth-toolkit==2.1.0
django-rest-auth==0.9.5
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-social-oauth2==1.2.1
ecdsa==0.18.0
idna==3.3
jwcrypto==1.3.1
oauthlib==3.2.0
Pillow==9.2.0
pyasn1==0.4.8
pycparser==2.21
PyJWT==2.4.0
python-jose==3.3.0
python3-openid==3.2.0
pytz==2022.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
wrapt==1.14.1
